feat - add command-line option to run in daemon mode (default now is non-deamon mode)
If there's no internet connection or any request errors, and there's
cached timings, they'll be used but the hijri date will continue as
"NULL".

In case of no cached timings, The program will retry the request again
every 10 minutes.

When connected and timings updated, in the non-daemon mode, timings are
formated and printed then the program ends. The damon mode is like
before, the timings will used to calculate the next prayer from current
time and print its name and countdown and all timings of the day in a
json format to be used with other programs like Waybar custom module.

This change of using daemon and non-daemon mode made no need for a
condition variable, just run the update worker anyway, the non-daemon
mode will return after that, and in the daemon mode it'll make sure
timings are updated before next prayer timing is calculated.